pyproject.toml
[tool.poetry]
name = "fx"
version = "0.1.0"
description = "Simple foreign exchange client-server application."
authors = ["Rodrigo Gryzinski <rogryza@rogryza.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
databases = {version = "^0.2.6", extras=["sqlite"]}
fastapi = "^0.54.1"
aiofiles = "^0.5.0"
uvicorn = "^0.11.3"
aiohttp = "^3.6.2"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
pylint = "^2.4.4"
pytest = "^5.4.1"
isort = "^4.3.21"
black = "^19.10b0"
sqlalchemy-stubs = "^0.3"
requests = "^2.23.0"
hypothesis = "^5.10.3"
pytest-asyncio = "^0.11.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
